Refactor code and improve UI components interaction
This commit includes several code clean-ups, particularly in the NewCoursePage.tsx and SectionList.tsx. The code was restructured in a more maintainable way and unnecessary properties were removed. Further, the commit introduces enhancements in the UI such as the addition of a dialog box for course creation in HomeHeader.tsx and better handling of list section items, enhancing UX.