Add functionality to delete question in Quiz Editor
Imported the `useSearchParams` and `CustomAlertDialog` components and added a delete operation for questions. A callback was created, `handleRemoveQuestion`, that triggers the question deletion process and it was tied to a new 'Remove question' button. The button resides inside a `CustomAlertDialog` that warns the user that the action is irreversible. Also, updated the styles for better UI and UX. Furthermore, a 'removeQuestion' mutation has been introduced in `question.tsx` to handle the deletion API call.