190 oll 370 redesign of the signup page in maker with the new design (#195)
* Add new background SVG for service-maker package

A new SVG file has been added to the public assets of the service-maker package to serve as its background. This new addition is intended to enhance the visual presentation of the package.

* Implement ProtectedRoute for course and editor pages

The ProtectedRoute wrapper was added to CourseManagerPage, QuizEditorPage, and LectureEditorPage in the routing definition. This change ensures that access to these pages is restricted to authorized users only. The import statement of ProtectedRoute has also been moved to the top of the file for better clarity.

* Refactor Login.tsx and improve user routing

Enhanced the code for Login.tsx by adding useCallback, useEffect, and making use of navigate from 'react-router-dom'. Refactored form submission logic and the API call structure. The onSubmit method now uses a callback method to redirect users, and useEffect checks if a user is already authenticated when visiting the page. Moreover, component styling was refactored to use styled-components.

* Add react-router-dom useNavigate hook to HomeSidePanel

The useNavigate hook from react-router-dom has been imported at the start of HomeSidePanel.tsx file. The previous location of its import, further down the file, has been removed. This reordering ensures better code organization and readability.

* Add Ollamy logo and styling to login form

The application logo "Ollamy" and its container have been added to the Login component. Additionally, several changes have been made to the container and form styling, including position, display, and dimensions adjustments to enhance the User Interface. Media query for wider screens is also included.

* Implement logout feature in HomeSidePanel

A logout action is added to userActions in the 'services/api/routes/user' file. This action is then imported into the HomeSidePanel component. A handleLogout function is defined to call the logout action when the logout button is clicked - afterward, the user is redirected to the home page.

* Refactor register feature and update navigation routes

The register page was completely refactored including form handling, UI and validation. Also, changes were made to navigation routes and cleaned up home navigation. Additionally, user service was updated; now on logout, user details are invalidated to keep data up to date.