Add edit functionality to course, section, and lesson (#198)
* Add edit functionality to course, section, and lesson

This commit introduces the ability to edit courses, sections, and lessons across the platform within the CourseManager module. The edit functionality is attached to a gear icon button in each respective resource. Furthermore, it adds update methods in the API routes to handle these actions.

* Update lesson editing and improve typescript hash resolution

Upgraded the lesson editing functionality in the CourseManager, allowing the updateLesson function to receive lessonId through moreOptions type. Also, updated the typescript hash resolution in the yarn.lock file to improve package management and security.