Add new Statistics page and refactor Home page structure (#221)
* Add new Statistics page and refactor Home page structure

Code structure of the Home page is refactored with new nested directories for better organization. A new Statistics page is added to the application, including its Header and Body components. The `@radix-ui/react-select` package version is also updated for better functionality of the select button.

* rm yarn.lock

* Add useMemo for total student calculation in StatisticsBody

A useMemo hook has been added to calculate the total number of students. This change improves performance by preventing unnecessary re-rendering of the component. Now, the total number of students enrolled is correctly reflected in the StatisticsBody.

* Refactor StatisticsBody component in Home Page

The changes include removing a redundant eslint-disable rule and redundant interface. Also reorganized import statements for better readability. Simplified the map function rendering CourseItems in the CourseSection for a cleaner code structure. Transitioned from constant to function declaration for better semantic understanding.