Fix build (#265)
* fx backend

* Fix typing and add new question type

Changed 'null' to 'undefined' for the 'time' property to better align with TypeScript typing. Added 'MULTIPLE_CHOICE' type to question creation and fixed syntax issues in SquareChoice.tsx methods.

* Add package-lock.json for dependency management

This commit adds the `package-lock.json` file to the repository, which helps ensure consistent installations of dependencies. This file captures the exact versions of all dependencies, aiding in reproducible builds and enhancing project stability.

* fix

---------

Co-authored-by: Nicolas HEUDE <nicolas.heude@epitech.eu>